[le_mesh] refactor
+ move ply loading logic to its own file so that le_mesh.cpp is a bit more
  focused.
