+++ modules/le_mesh/CMakeLists.txt
@@ -1,8 +1,9 @@
 set (TARGET le_mesh)
 
-set (SOURCES "le_mesh.cpp")
-set (SOURCES ${SOURCES} "le_mesh.h")
-set (SOURCES ${SOURCES} "le_mesh_types.h")
+set (SOURCES "le_mesh.h" )
+set (SOURCES ${SOURCES} "le_mesh.cpp" )
+set (SOURCES ${SOURCES} "le_mesh_types.h" )
+set (SOURCES ${SOURCES} "le_mesh_ply.cpp" )
 
 if (${PLUGINS_DYNAMIC})
     add_library(${TARGET} SHARED ${SOURCES})
